services/dramatiq-classifiedad/requirements.txt
async-timeout==4.0.3
certifi==2023.11.17
cffi==1.16.0
circuitbreaker==1.4.0
cryptography==41.0.7
dramatiq==1.15.0
gevent==23.9.1
greenlet==3.0.3
mysql-connector-python==8.2.0
oci==2.118.0
prometheus-client==0.19.0
protobuf==4.21.12
pycparser==2.21
pyOpenSSL==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
watchdog==3.0.0
watchdog-gevent==0.1.1
zope.event==5.0
zope.interface==6.1
